Move assignments to correct package so that hints are shown
