+++ webgoat-lessons/sql-injection/src/main/java/org/owasp/webgoat/plugin/advanced/SqlInjectionLesson6b.java
@@ -1,8 +1,7 @@
 
-package org.owasp.webgoat.plugin.introduction;
+package org.owasp.webgoat.plugin.advanced;
 
 import org.owasp.webgoat.assignments.AssignmentEndpoint;
-import org.owasp.webgoat.assignments.AssignmentHints;
 import org.owasp.webgoat.assignments.AssignmentPath;
 import org.owasp.webgoat.assignments.AttackResult;
 import org.owasp.webgoat.session.DatabaseUtilities;
@@ -12,7 +11,10 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 
 import java.io.IOException;
-import java.sql.*;
+import java.sql.Connection;
+import java.sql.ResultSet;
+import java.sql.SQLException;
+import java.sql.Statement;
 
 
 /***************************************************************************************************
@@ -45,53 +47,42 @@ import java.sql.*;
  * @author Bruce Mayhew <a href="http://code.google.com/p/webgoat">WebGoat</a>
  * @created October 28, 2003
  */
-@AssignmentPath("/SqlInjection/attack6a")
-@AssignmentHints(value = {"SqlStringInjectionHint5", "SqlStringInjectionHint6", "SqlStringInjectionHint7"})
-public class SqlInjectionLesson6a extends AssignmentEndpoint {
+@AssignmentPath("/SqlInjection/attack6b")
+public class SqlInjectionLesson6b extends AssignmentEndpoint {
 
     @RequestMapping(method = RequestMethod.POST)
-    public
     @ResponseBody
-    AttackResult completed(@RequestParam String userid_6a) throws IOException {
-        return injectableQuery(userid_6a);
-        // The answer: Smith' union select userid,user_name, password,cookie,cookie, cookie,userid from user_system_data --
-
+    public AttackResult completed(@RequestParam String userid_6b) throws IOException {
+        if (userid_6b.toString().equals(getPassword())) {
+            return trackProgress(success().build());
+        } else {
+            return trackProgress(failed().build());
+        }
     }
 
-    protected AttackResult injectableQuery(String accountName) {
+    protected String getPassword() {
+
+        String password = "dave";
         try {
             Connection connection = DatabaseUtilities.getConnection(getWebSession());
-            String query = "SELECT * FROM user_data WHERE last_name = '" + accountName + "'";
+            String query = "SELECT password FROM user_system_data WHERE user_name = 'dave'";
 
             try {
                 Statement statement = connection.createStatement(ResultSet.TYPE_SCROLL_INSENSITIVE,
                         ResultSet.CONCUR_READ_ONLY);
                 ResultSet results = statement.executeQuery(query);
 
-                if ((results != null) && (results.first())) {
-                    ResultSetMetaData resultsMetaData = results.getMetaData();
-                    StringBuffer output = new StringBuffer();
-
-                    output.append(SqlInjectionLesson5a.writeTable(results, resultsMetaData));
-                    results.last();
-
-                    // If they get back more than one user they succeeded
-                    if (results.getRow() >= 5) {
-                        return trackProgress(success().feedback("sql-injection.6a.success").feedbackArgs(output.toString()).build());
-                    } else {
-                        return trackProgress(failed().output(output.toString()).build());
-                    }
-
-                } else {
-                    return trackProgress(failed().feedback("sql-injection.6a.no.results").build());
-
+                if ((results != null) && (results.first() == true)) {
+                    password = results.getString("password");
                 }
             } catch (SQLException sqle) {
-                return trackProgress(failed().output(sqle.getMessage()).build());
+                sqle.printStackTrace();
+                // do nothing
             }
         } catch (Exception e) {
             e.printStackTrace();
-            return trackProgress(failed().output(this.getClass().getName() + " : " + e.getMessage()).build());
+            // do nothing
         }
+        return (password);
     }
 }
